+++ container_service_extension/broker_selector.py
@@ -2,8 +2,6 @@
 # Copyright (c) 2017 VMware, Inc. All Rights Reserved.
 # SPDX-License-Identifier: BSD-2-Clause
 
-from vcd_cli.profiles import Profiles
-
 from container_service_extension import utils
 from container_service_extension.broker import DefaultBroker
 from container_service_extension.logger import SERVER_LOGGER as LOGGER
@@ -13,9 +11,13 @@ from container_service_extension.pksbroker import PKSBroker
 
 def get_new_broker(headers, request_body):
 
-    profiles = Profiles.load()
-    ovdc_name = profiles.get('vdc_in_use')
-    org_name = profiles.get('org_in_use')
+    config = utils.get_server_runtime_config()
+    tenant_client, session = utils.connect_vcd_user_via_token(
+        vcd_uri=config['vcd']['host'],
+        headers=headers,
+        verify_ssl_certs=config['vcd']['verify'])
+    ovdc_name = request_body.get('vdc') if request_body else None
+    org_name = session.get('org')
     LOGGER.debug(f"org_name={org_name};vdc_name=\'{ovdc_name}\'")
 
     """
@@ -35,4 +37,7 @@ def get_new_broker(headers, request_body):
         else:
             return DefaultBroker(headers, request_body)
     else:
+        # TODO() - This call should be based on a boolean flag
+        # Specify flag in config file whether to have default
+        # handling is required for missing ovdc or org.
         return DefaultBroker(headers, request_body)
